requirements.txt
# Production requirements for deployment
# FastAPI and web server
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Database
pymongo==4.6.1
motor==3.3.2

# OpenAI LLM
openai==1.12.0

# Data processing and validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
requests==2.31.0
numpy==1.26.3
geopy==2.4.0

# Logging
loguru==0.7.2

# Environment variables
python-dotenv==1.0.0

# Additional utilities for Python 3.12 compatibility
typing-extensions==4.9.0

# Production optimizations
gunicorn==21.2.0
prometheus-client==0.19.0